Split reducer from store creation and control store creation
