+++ src/components/chat-view-container/chat-view-container.test.tsx
@@ -1,4 +1,4 @@
-import { RootState } from '../../store';
+import { RootState } from '../../store/reducer';
 
 import { shallow } from 'enzyme';
