+++ src/components/chat-view-container/chat-view-container.tsx
@@ -1,6 +1,6 @@
 import React, { RefObject } from 'react';
 import classNames from 'classnames';
-import { RootState } from '../../store';
+import { RootState } from '../../store/reducer';
 
 import { connectContainer } from '../../store/redux-container';
 import {
