+++ src/components/notification/list/container.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { connectContainer } from '../../../store/redux-container';
-import { RootState } from '../../../store';
+import { RootState } from '../../../store/reducer';
 import { fetch as fetchNotifications, denormalizeNotifications } from '../../../store/notifications';
 import { denormalize as denormalizeChannel } from '../../../store/channels';
