+++ src/store/index.ts
@@ -1,20 +1,8 @@
 import { configureStore } from '@reduxjs/toolkit';
 import createSagaMiddleware from 'redux-saga';
-import { combineReducers } from 'redux';
 
 import { rootSaga } from './saga';
-
-import { reducer as layout } from './layout';
-import { reducer as channelsList } from './channels-list';
-import { reducer as web3 } from './web3';
-import { reducer as zns } from './zns';
-import { reducer as theme } from './theme';
-import { reducer as apps } from './apps';
-import { reducer as normalized } from './normalized';
-import { reducer as authentication } from './authentication';
-import { reducer as chat } from './chat';
-import { reducer as notificationsList } from './notifications';
-import { reducer as createConversation } from './create-conversation';
+import { rootReducer } from './reducer';
 
 const sagaMiddleware = createSagaMiddleware({
   onError: (e) => {
@@ -22,25 +10,11 @@ const sagaMiddleware = createSagaMiddleware({
   },
 });
 
-export const rootReducer = combineReducers({
-  layout,
-  channelsList,
-  web3,
-  zns,
-  theme,
-  apps,
-  normalized,
-  authentication,
-  chat,
-  notificationsList,
-  createConversation,
-});
-
-export type RootState = ReturnType<typeof rootReducer>;
-
 export const store = configureStore({
   reducer: rootReducer,
   middleware: (defaults) => defaults({ thunk: false }).concat(sagaMiddleware),
 });
 
-sagaMiddleware.run(rootSaga);
+export function runSagas() {
+  sagaMiddleware.run(rootSaga);
+}
