+++ src/store/notifications/saga.test.ts
@@ -3,7 +3,7 @@ import { expectSaga, testSaga } from 'redux-saga-test-plan';
 import * as matchers from 'redux-saga-test-plan/matchers';
 
 import { AsyncListStatus } from '../normalized';
-import { rootReducer } from '..';
+import { rootReducer } from '../reducer';
 
 import {
   addNotification,
